requirements.txt
flask
opencv-python
dlib
numpy
pillow